+++ ETrainerWEB/Models/Dish.cs
@@ -1,6 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Text.Json.Serialization;
 
 namespace ETrainerWEB.Models
 {
@@ -13,11 +11,10 @@ namespace ETrainerWEB.Models
         }
         public int Id { set; get; }
         public string Name { set; get; }
-        public float PortionWeight { set; get; }
-        public float CaloricityPerGram { set; get; }
+        public double PortionWeight { set; get; }
+        public double CaloricityPerGram { set; get; }
         public User User { set; get; }
         public ICollection<MealsDishes> MealsDishes { get; set; }
         public ICollection<DishesIngredients> DishesIngredients { get; set; }
-
     }
 }
