+++ ETrainerWEB/Models/Ingredient.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Text.Json.Serialization;
 
 namespace ETrainerWEB.Models
 {
@@ -7,8 +6,8 @@ namespace ETrainerWEB.Models
     {
         public int Id { set; get; }
         public string Name { set; get; }
-        public float PortionWeight { set; get; }
-        public float CaloricityPerGram { set; get; }
+        public double PortionWeight { set; get; }
+        public double CaloricityPerGram { set; get; }
         public User User { set; get; }
         public ICollection<DishesIngredients> DishesIngredients { get; set; }
     }
